+++ services/director-v2/src/simcore_service_director_v2/modules/db/repositories/comp_tasks.py
@@ -2,7 +2,6 @@ import logging
 from datetime import datetime
 from typing import Dict
 
-import networkx as nx
 import sqlalchemy as sa
 from sqlalchemy.dialects.postgresql import insert
 
@@ -15,36 +14,16 @@ from models_library.services import (
     ServiceType,
 )
 
-from ....models.domains.comp_pipelines import CompPipelineAtDB
 from ....models.domains.comp_tasks import CompTaskAtDB, Image, NodeSchema
 from ....models.domains.projects import ProjectAtDB
 from ....models.schemas.services import NodeRequirement, ServiceExtras
 from ....utils.computations import to_node_class
 from ....utils.logging_utils import log_decorator
 from ...director_v0 import DirectorV0Client
-from ..tables import NodeClass, StateType, comp_pipeline, comp_tasks
+from ..tables import NodeClass, StateType, comp_tasks
 from ._base import BaseRepository
 
 logger = logging.getLogger(__name__)
-
-
-class CompPipelinesRepository(BaseRepository):
-    @log_decorator(logger=logger)
-    async def publish_pipeline(
-        self, project_id: ProjectID, dag_graph: nx.DiGraph
-    ) -> None:
-
-        pipeline_at_db = CompPipelineAtDB(
-            project_id=project_id,
-            dag_adjacency_list=nx.to_dict_of_lists(dag_graph),
-            state=RunningState.PUBLISHED,
-        )
-        insert_stmt = insert(comp_pipeline).values(**pipeline_at_db.dict(by_alias=True))
-        on_update_stmt = insert_stmt.on_conflict_do_update(
-            index_elements=[comp_pipeline.c.project_id],
-            set_=pipeline_at_db.dict(by_alias=True, exclude_unset=True),
-        )
-        await self.connection.execute(on_update_stmt)
 
 
 @log_decorator(logger=logger)
@@ -174,4 +153,10 @@ class CompTasksRepository(BaseRepository):
                 )
             )
             .values(state=StateType.ABORTED)
+        )
+
+    @log_decorator(logger=logger)
+    async def delete_tasks_from_project(self, project: ProjectAtDB) -> None:
+        await self.connection.execute(
+            sa.delete(comp_tasks).where(comp_tasks.c.project_id == str(project.uuid))
         )
